requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
requests==2.31.0
numpy==1.26.4
matplotlib==3.8.2
pytest==8.0.2
pandas==2.1.4
python-binance==1.0.21
tabulate==0.9.0